% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{AggPopData}
\alias{AggPopData}
\title{Read and prepare aggregated population input.}
\usage{
AggPopData(wdir, StatusDate, AggregatedPopulationFile)
}
\arguments{
\item{wdir}{Working directory.}

\item{StatusDate}{Date of status calculation.}

\item{AggregatedPopulationFile}{Aggregated input ;-separated population file}
}
\value{
AggPop
}
\description{
Read and prepare aggregated population input.
}
