% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{AggVacData}
\alias{AggVacData}
\title{Read and prepare aggregated vaccination input.}
\usage{
AggVacData(wdir, StatusDate, AggregatedMCVFile)
}
\arguments{
\item{wdir}{Working directory.}

\item{StatusDate}{Date of status calculation.}

\item{AggregatedMCVFile}{Aggregated input ;-separated vaccination file}
}
\value{
AggVac
}
\description{
Read and prepare aggregated vaccination input.
}
