% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractDKData.R
\name{DKIndPop}
\alias{DKIndPop}
\title{Extract Danish individuel population file.}
\usage{
DKIndPop(ExtractDate, wdir)
}
\arguments{
\item{ExtractDate}{Date of extraction.}

\item{wdir}{Working directory.}
}
\value{
;-seperated file
}
\description{
Extract Danish individuel population file.
}
