% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractDKData.R
\name{DKIndVac}
\alias{DKIndVac}
\title{Extract Danish individuel vaccination file.}
\usage{
DKIndVac(ExtractDate, wdir)
}
\arguments{
\item{ExtractDate}{Date of extraction.}

\item{wdir}{Working directory.}
}
\value{
;-seperated file
}
\description{
Extract Danish individuel vaccination file.
}
