% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{IndPopData}
\alias{IndPopData}
\title{Read and prepare individual population input.}
\usage{
IndPopData(wdir, StatusDate, IndividuelPopulationFile)
}
\arguments{
\item{wdir}{Working directory.}

\item{StatusDate}{Date of status calculation.}

\item{IndividuelPopulationFile}{Individual input ;-separated population file}
}
\value{
IndPop
}
\description{
Read and prepare individual population input.
}
