% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{IndVacData}
\alias{IndVacData}
\title{Read and prepare individual vaccination input.}
\usage{
IndVacData(wdir, StatusDate, IndividualMCVFile)
}
\arguments{
\item{wdir}{Working directory.}

\item{StatusDate}{Date of status calculation.}

\item{IndividualMCVFile}{Individual input ;-separated vaccination file}
}
\value{
IndVac
}
\description{
Read and prepare individual vaccination input.
}
