% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{ShowCov}
\alias{ShowCov}
\title{Show coverage graphs and maps on a Shiny dashboard.}
\usage{
ShowCov(wdir)
}
\arguments{
\item{wdir}{Path to directory where data are (directory with Status_yyyymmdd subdirectories)}
}
\value{
Open local Shiny dashboard. Exit by closing dashboad window.
}
\description{
Show coverage graphs and maps on a Shiny dashboard.
}
