% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{TimeMonth}
\alias{TimeMonth}
\title{Calculate time-difference in month.}
\usage{
TimeMonth(StartDate, EndDate)
}
\arguments{
\item{StartDate}{Starting date}

\item{EndDate}{End date}
}
\value{
time-diff in month
}
\description{
Calculate time-difference in month.
}
