% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{VacCovAge}
\alias{VacCovAge}
\title{Coverage graphs by age of vaccination.}
\usage{
VacCovAge(wdir, StatusDate, NUTS, dose, save = FALSE)
}
\arguments{
\item{wdir}{Working directory.}

\item{StatusDate}{Date of status calculation.}

\item{NUTS}{NUTS Id}

\item{dose}{First or second dose}

\item{save}{Should the graphs be saved to disk or returned}
}
\value{
Coverage graph on \code{NUTS} by age at vaccination
}
\description{
Coverage graphs by age of vaccination.
}
