% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{VacCovBC}
\alias{VacCovBC}
\title{Coverage graphs by birth cohort.}
\usage{
VacCovBC(wdir, StatusDate, NUTS, dose, save = FALSE)
}
\arguments{
\item{wdir}{Working directory.}

\item{StatusDate}{Date of status calculation.}

\item{NUTS}{NUTS Id}

\item{dose}{First or second dose}

\item{save}{Should the graphs be saved to disk or returned}
}
\value{
Coverage graph on \code{NUTS} by Birth cohort
}
\description{
Coverage graphs by birth cohort.
}
