% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{VacCovGraphs}
\alias{VacCovGraphs}
\title{Create coverage graphs at NUTS-level.}
\usage{
VacCovGraphs(wdir, StatusDate, NUTSlevel)
}
\arguments{
\item{wdir}{Working directory.}

\item{StatusDate}{Date of status calculation.}

\item{NUTSlevel}{NUTS level}
}
\value{
Graphs
}
\description{
Create coverage graphs at NUTS-level.
}
