% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{MapCov}
\alias{MapCov}
\title{Create coverage map at NUTS-level by Birth Cohort.}
\usage{
MapCov(wdir, StatusDate, BirthCohort, dose)
}
\arguments{
\item{wdir}{Working directory.}

\item{StatusDate}{Date of status calculation.}

\item{BirthCohort}{Birth year of cohort}

\item{dose}{First or second dose}
}
\value{
map
}
\description{
Create coverage map at NUTS-level by Birth Cohort.
}
